pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mancala-solver"
version = "0.1.0"
description = "Strong solver for Kalah/Mancala variants"
authors = [{name = "Jacob Ellison"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
cloud = [
    "psycopg2-binary>=2.9.0",
    "google-cloud-bigtable>=2.0.0",
    "google-cloud-storage>=2.0.0",
    "pyarrow>=12.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "memory-profiler>=0.60.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
